Cleaning up code and removing unused files
